cleaner pre_init/post_init/etc configuration
