+++ pyterrier/java/__init__.py
@@ -1,5 +1,5 @@
 from pyterrier.java import mavenresolver
-from pyterrier.java._init import init, started, parallel_init, parallel_init_args, required, required_raise, before_init, autoclass, cast, legacy_init
+from pyterrier.java._init import init, started, parallel_init, parallel_init_args, required, required_raise, before_init, autoclass, cast, legacy_init, JavaInitializer, CoreInit
 from pyterrier.java import config
 from pyterrier.java.config import configure
 from pyterrier.java.utils import redirect_stdouterr, bytebuffer_to_array, JavaClasses, J, add_jar, add_package, set_memory_limit, set_redirect_io, add_option, set_log_level
@@ -8,5 +8,5 @@ __all__ = [
     'add_jar', 'add_option', 'add_package', 'autoclass', 'before_init', 'bytebuffer_to_array', 'cast', 'config',
     'configure', 'init', 'J', 'JavaClasses', 'mavenresolver', 'parallel_init', 'parallel_init_args',
     'redirect_stdouterr', 'required', 'required_raise', 'set_log_level', 'set_memory_limit', 'set_redirect_io',
-    'started', 'legacy_init',
+    'started', 'legacy_init', 'JavaInitializer', 'CoreInit',
 ]
