+++ pyterrier/terrier/index.py
@@ -13,7 +13,7 @@ import math
 from warnings import warn
 from deprecated import deprecated
 from collections import deque
-from typing import List, Dict, Union, Any
+from typing import List, Dict, Union
 import more_itertools
 import pyterrier as pt
 from pyterrier.terrier.stemmer import TerrierStemmer
@@ -21,7 +21,7 @@ from pyterrier.terrier.tokeniser import TerrierTokeniser
 from pyterrier.terrier.stopwords import TerrierStopwords
 
 
-# These classes are only defined after pt.java.init()
+# These classes are only defined after pt.java.init() in pyterrier.terrier.java._post_init
 DocListIterator = None
 PythonListIterator = None
 FlatJSONDocumentIterator = None
@@ -36,214 +36,6 @@ class IterDictIndexerBase(pt.Indexer):
 
 # lastdoc ensures that a Document instance from a Collection is not GCd before Java has used it.
 lastdoc = None
-
-def _java_post_init(jnius):
-    global DocListIterator
-    global PythonListIterator
-    global FlatJSONDocumentIterator
-    global TQDMCollection
-    global TQDMSizeCollection
-
-    @pt.java.required
-    class DocListIterator(jnius.PythonJavaClass):
-        __javainterfaces__ = [
-            'java/util/Iterator',
-        ]
-
-        def __init__(self, pyiterator):
-            self.pyiterator = pyiterator
-            self.hasnext = True
-            self.lastdoc = None
-            self.tr57 = not pt.terrier.check_version("5.8")
-
-        @staticmethod 
-        def pyDictToMap(a_dict): #returns Map<String,String>
-            rtr = pt.java.J.HashMap()
-            for k,v in a_dict.items():
-                rtr.put(k, v)
-            return rtr
-
-        def pyDictToMapEntry(self,doc_dict : Dict[str,Any]): #returns Map.Entry<Map<String,String>, DocumentPostingList>>
-            dpl = pt.terrier.J.DocumentPostingList()
-            # this works around a bug in the counting of doc lengths in Tr 5.7
-            if self.tr57:
-                for t, tf in doc_dict["toks"].items():
-                    for i in range(int(tf)):
-                        dpl.insert(t)
-            else: # this code for 5.8 onwards
-                for t, tf in doc_dict["toks"].items():
-                    dpl.insert(int(tf), t)
-            
-            # we cant make the toks column into the metaindex as it isnt a string. remove it.
-            del doc_dict["toks"]
-            return pt.terrier.J.MapEntry(DocListIterator.pyDictToMap(doc_dict), dpl)
-
-        @jnius.java_method('()Z')
-        def hasNext(self):
-            return self.hasnext
-
-        @jnius.java_method('()Ljava/lang/Object;')
-        def next(self):
-            try:
-                doc_dict = next(self.pyiterator)
-            except StopIteration as se:
-                self.hasnext = False
-                # terrier will ignore a null return from an iterator
-                return None
-            # keep this around to prevent being GCd before Java can read it
-            self.lastdoc = self.pyDictToMapEntry(doc_dict)
-            return self.lastdoc
-            
-    @pt.java.required
-    class TQDMCollection(jnius.PythonJavaClass):
-        __javainterfaces__ = ['org/terrier/indexing/Collection']
-
-        def __init__(self, collection):
-            super(TQDMCollection, self).__init__()
-            assert isinstance(collection, pt.terrier.J.MultiDocumentFileCollection)
-            self.collection = collection
-            size = self.collection.FilesToProcess.size()
-            self.pbar = pt.tqdm(total=size, unit="files")
-            self.last = -1
-        
-        @jnius.java_method('()Z')
-        def nextDocument(self):
-            rtr = self.collection.nextDocument()
-            filenum = self.collection.FileNumber
-            if filenum > self.last:
-                self.pbar.update(filenum - self.last)
-                self.last = filenum
-            return rtr
-
-        @jnius.java_method('()V')
-        def reset(self):
-            self.pbar.reset()
-            self.collection.reset()
-
-        @jnius.java_method('()V')
-        def close(self):
-            self.pbar.close()
-            self.collection.close()
-
-        @jnius.java_method('()Z')
-        def endOfCollection(self):
-            return self.collection.endOfCollection()
-
-        @jnius.java_method('()Lorg/terrier/indexing/Document;')
-        def getDocument(self):
-            global lastdoc
-            lastdoc = self.collection.getDocument()
-            return lastdoc
-
-    class PythonListIterator(jnius.PythonJavaClass):
-        __javainterfaces__ = ['java/util/Iterator']
-
-        def __init__(self, text, meta, convertFn, len=None, index=0):
-            super(PythonListIterator, self).__init__()
-            self.text = text
-            self.meta = meta
-            self.index = index
-            self.convertFn = convertFn
-            if len is None:
-                self.len = len(self.text)
-            else:
-                self.len = len
-
-        @jnius.java_method('()V')
-        def remove():
-            # 1
-            pass
-
-        @jnius.java_method('(Ljava/util/function/Consumer;)V')
-        def forEachRemaining(action):
-            # 1
-            pass
-
-        @jnius.java_method('()Z')
-        def hasNext(self):
-            return self.index < self.len
-
-        @jnius.java_method('()Ljava/lang/Object;')
-        def next(self):
-            text = self.text[self.index]
-            meta = self.meta.__next__()
-            self.index += 1
-            global lastdoc
-            if self.convertFn is not None:
-                lastdoc = self.convertFn(text, meta)
-            else:
-                lastdoc = [text, meta]
-            return lastdoc
-
-    @pt.java.required
-    class FlatJSONDocumentIterator(jnius.PythonJavaClass):
-        __javainterfaces__ = ['java/util/Iterator']
-
-        def __init__(self, it):
-            super(FlatJSONDocumentIterator, self).__init__()
-            self._it = it
-            # easiest way to support hasNext is just to start consuming right away, I think
-            self._next = next(self._it, StopIteration)
-
-        @jnius.java_method('()V')
-        def remove():
-            # 1
-            pass
-
-        @jnius.java_method('(Ljava/util/function/Consumer;)V')
-        def forEachRemaining(action):
-            # 1
-            pass
-
-        @jnius.java_method('()Z')
-        def hasNext(self):
-            return self._next is not StopIteration
-
-        @jnius.java_method('()Ljava/lang/Object;')
-        def next(self):
-            result = self._next
-            self._next = next(self._it, StopIteration)
-            if result is not StopIteration:
-                global lastdoc
-                lastdoc = pt.terrier.J.FlatJSONDocument(json.dumps(result))
-                return lastdoc
-            return None
-
-    class TQDMSizeCollection(jnius.PythonJavaClass):
-        __javainterfaces__ = ['org/terrier/indexing/Collection']
-
-        def __init__(self, collection, total):
-            super(TQDMSizeCollection, self).__init__()
-            self.collection = collection
-            self.pbar = pt.tqdm(total=total, unit="documents")
-        
-        @jnius.java_method('()Z')
-        def nextDocument(self):
-            rtr = self.collection.nextDocument()
-            self.pbar.update()
-            return rtr
-
-        @jnius.java_method('()V')
-        def reset(self):
-            self.pbar.reset()
-            self.collection.reset()
-
-        @jnius.java_method('()V')
-        def close(self):
-            self.pbar.close()
-            self.collection.close()
-
-        @jnius.java_method('()Z')
-        def endOfCollection(self):
-            return self.collection.endOfCollection()
-
-        @jnius.java_method('()Lorg/terrier/indexing/Document;')
-        def getDocument(self):
-            global lastdoc
-            lastdoc = self.collection.getDocument()
-            return lastdoc
-
-
 
 type_to_class = {
     'trec' : 'org.terrier.indexing.TRECCollection',
